fixed jittering when colliding with walls
fixed wall jitter bug (hehehe)
if we change the angle of inclines we will need to modify the variable used by the angle, 45 degrees is equal to 0.5 in the float, so if its 90 degrees it would be 1 kinda:
basically it works like this right, if you're on a 45 degree angle, your angle is 0.5 in the y axis and 0.5 in the direction the angle is, unless its diagonal. assuming its straight with the z axis for example, the normal would be: (0,0.5,0.5) now we only need to worry about the Y value for this.

if we change the angle that mango can go up ramps we will need to see the new y angle from that using this equation:
x/90 = y

x should be the angle in degrees on the y axis that the ramp is, y is the new value we need to subtract and add to the rampDown and rampUp vars respectively